requirements.txt
dj-database-url==0.4.1
Django==1.9.6
requests==2.10.0
pycodestyle==1.8.0.dev0
psycopg2==2.6.1
python-social-auth==0.2.19
gunicorn==19.6.0
whitenoise==3.1
pygithub==1.26.0
django-rq==0.9.1